# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/me/devices/push-token": {
      "put": {
        "operationId": "DevicesController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Register this device for push notifications (this session)",
        "tags": [
          "me · devices"
        ]
      },
      "delete": {
        "operationId": "DevicesController_remove_v1",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Stop push notifications to this device",
        "tags": [
          "me · devices"
        ]
      }
    },
    "/v1/me/notifications": {
      "get": {
        "operationId": "NotificationsController_list_v1",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 30,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPageDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "In-app notifications, newest first, with the unread count",
        "tags": [
          "me · notifications"
        ]
      }
    },
    "/v1/me/notifications/read": {
      "post": {
        "operationId": "NotificationsController_read_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarkNotificationsReadDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mark notifications read (up to one, or all)",
        "tags": [
          "me · notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Sajha API",
    "description": "REST API for the Sajha mobile app, admin panel and website.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PushTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "FCM registration token"
          },
          "platform": {
            "type": "string",
            "enum": [
              "android",
              "ios"
            ]
          }
        },
        "required": [
          "token",
          "platform"
        ]
      },
      "NotificationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "example": "booking.requested"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "bookingId": {
            "type": "string",
            "nullable": true,
            "format": "uuid",
            "description": "Open this booking when tapped"
          },
          "readAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "body",
          "createdAt"
        ]
      },
      "NotificationPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Newest first",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          },
          "unread": {
            "type": "number",
            "description": "Unread notifications in total"
          }
        },
        "required": [
          "items",
          "unread"
        ]
      },
      "MarkNotificationsReadDto": {
        "type": "object",
        "properties": {
          "upTo": {
            "type": "string",
            "format": "uuid",
            "description": "Everything up to and including this one; omit to mark all read"
          }
        }
      },
...
    }
  }
}